extracted some of the components
