+++ nextjs-frontend/README.md
@@ -21,3 +21,7 @@ npm install leaflet-defaulticon-compatibility
 npm i react-div-100vh
 npm install -D @tailwindcss/typography
 ```
+
+## Todo
+
+- Add spinner background and make it load
