+++ nextjs-frontend/styles/globals.css
@@ -19,4 +19,36 @@
   background-position: 0 0;
   background-size: 65px 20px;
 }
+
+/* .overlay {
+  background-color: rgba(0, 0, 0, 0.5);
+  height: 100%;
+  width: 100%;
+  position: absolute;
+  left: 0px;
+  right: 0px;
+  z-index: 1200; 
+} */
+
+.loader {
+  border: 16px solid #f3f3f3; 
+  border-top: 16px solid #3498db; 
+  border-radius: 50%;
+  width: 100px;
+  height: 100px;
+  animation: spin 2s linear infinite;
+  z-index: 1600;
+  align-self: center;
+  margin: auto;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
 @tailwind utilities;
